# src/generated/epr-flow-control-api/model/reportVersionsResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Matrix Reports API
 * API for managing energy reports with pagination and filters
 * OpenAPI spec version: 1.0
 */

export interface ReportVersionsResponseDto {
  /** Correlation ID of the report */
  correlationId: string;
  /** List of versions available for the report */
  versions: string[];
}