# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reports": {
      "get": {
        "description": "Returns a list of reports.",
        "operationId": "ReportsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 0,
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for UC, meter point, or nickname (case-insensitive contains)",
            "schema": {
              "example": "SP",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status (can be multiple). Accepts array or comma-separated values. Example: ?status=SENT&status=READY_TO_SEND or ?status=SENT,READY_TO_SEND",
            "schema": {
              "example": [
                "READY_TO_GENERATE",
                "SENT"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "READY_TO_GENERATE",
                  "GENERATING_REPORT",
                  "PENDING_INFORMATION",
                  "NEED_ANALYSIS",
                  "PROCESSING_ERROR",
                  "READY_TO_SEND",
                  "QUEUED_FOR_SENDING",
                  "SENT",
                  "SENDING_ERROR"
                ]
              }
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for filtering reports",
            "schema": {
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for filtering reports",
            "schema": {
              "example": "2026-02-01",
              "type": "string"
            }
          },
          {
            "name": "economicGroup",
            "required": false,
            "in": "query",
            "description": "Filter by economic group",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Sort by field",
            "schema": {
              "default": "referenceDate",
              "example": "referenceDate",
              "type": "string",
              "enum": [
                "referenceDate",
                "codeUc",
                "nickname",
                "economicGroup",
                "reportStatus",
                "updatedAt"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved reports list"
          }
        },
        "summary": "List Reports",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/reports/{correlationId}/versions": {
      "get": {
        "description": "Returns the mocked list of versions for a specific report correlationId.",
        "operationId": "ReportsController_versions",
        "parameters": [
          {
            "name": "correlationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved report versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportVersionsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Report not found"
          }
        },
        "summary": "List report versions by correlationId",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/reports/{correlationId}/generate": {
      "post": {
        "description": "Triggers the generation of a report for a specific correlationId and sends it to the processing queue.",
        "operationId": "ReportsController_generateByCorrelationId",
        "parameters": [
          {
            "name": "correlationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Report generation initiated successfully"
          }
        },
        "summary": "Generate Report by correlationId",
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Matrix Reports API",
    "description": "API for managing energy reports with pagination and filters",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ReportVersionsResponseDto": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string",
            "description": "Correlation ID of the report",
            "example": "mock-correlation-000001"
          },
          "versions": {
            "description": "List of versions available for the report",
            "example": [
              "1.0.0"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "correlationId",
          "versions"
        ]
      },
...
    }
  }
}